+++ models.py
@@ -20,8 +20,8 @@ class User(DB.Model):
     id = Column(Integer, autoincrement=True, primary_key=True)
     active = Column(Boolean, default=True)
     username = Column(String(20), unique=True)
-    email = Column(String(200))
-    confirm_mail = Column(Boolean, default=True)
+    email = Column(String(200), unique=True)
+    confirm_mail = Column(Boolean, default=False)
     password = Column(String(200), default='')
     modderator = Column(Boolean, default=False)
     admin = Column(Boolean, default=False)
