+++ requirements/common.txt
@@ -1,6 +1,7 @@
 Flask==0.12.2
 Flask-Login==0.4.1
 Flask-SQLAlchemy==2.3.2
+Flask-Mail==0.9.1
 Jinja2==2.10
 SQLAlchemy==1.2.5
 Werkzeug==0.12.2
