+++ templates/user.html
@@ -21,6 +21,7 @@
           {% endif %}
         </h2>
         <h4>{{user.email}}</h4>
+        <h4>{{user.confirm_mail}}</h4>
     </div>
     <div class="col-md-4">
          {%if user.username==current_user.username%}
